Observables

Observables

Atoms.Observable — Type.

Observable

Represents a physical quantity that can be observed, which is calculated as the matrix element of an operator between two configurations. All physical observables are real.

source

Atoms.Observable — Method.

Observable(operator, atom, overlaps, integrals, integral_map, selector
           [; double_counted=false])

Construct an observable corresponding the operator acting on atom; if double_counted, only return those terms that would be double-counted, otherwise return the normal observable equations. overlaps is a list of non-orthogonal, integrals a list of common integrals, and integral_map is a mapping from symbolic integrals to OrbitalIntegrals. selector selects those orbitals not modelled by the potential, e.g. all orbitals in case of a nuclear potential, fewer in case of a pseudopotential. double_counted can be used to only return those terms of the sum that would be double-counted if simply summing over orbital contributions (applicable to two-body operators only). Half the result of the double-counted term is then subtracted the sum over orbital contributions; this makes the expression slightly more symmetric in orbital space than if the sum is derived avoiding double-counting.

source

Atoms.observe! — Function.

observe!(A::M, o::Observable)

Compute the observable o between all configurations and store the results as matrix elements of A.

source

observe!(A::M, o::Observable, atom::Atom)

Compute the observable o between all configurations, store the results as matrix elements of A, and finally contract with respect to the mixing coefficients of atom and return the result as a scalar.

source