Orbital equations

Orbital equations

Hamiltonian

Atoms.OrbitalHamiltonianTerm — Type.

OrbitalHamiltonianTerm(i, j, coeff, A, integrals)

Represents a term in the orbital Hamiltonian arising from a variation of the energy expressions between configurations i and j in the multi-configurational expansion. coeff is the numeric coefficient, A is the operator acting on the orbital, and integrals is a vector of OrbitalIntegrals arising from the presence of non-orthogonal orbitals and whose values should be multiplied to form the overall coefficient.

source

Atoms.coefficient — Function.

coefficient(term::OrbitalHamiltonianTerm)

Return the multiplicative coefficient pertaining to term, excluding the conj(c_i)*c_j mixing coefficients, due to the configuration-interaction.

source

coefficient(term::OrbitalHamiltonianTerm, c::Vector)

Return the multiplicative coefficient pertaining to term, including the conj(c_i)*c_j mixing coefficients, due to the configuration-interaction.

source

Atoms.OrbitalHamiltonian — Type.

OrbitalHamiltonian(R, terms, mix_coeffs, projector, orbital)

The Hamiltonian for orbital is constructed from a radial basis R, a set of OrbitalHamiltonianTerm terms that describe the various interactions between orbitals, mix_coeffs which are the mixing coefficents for the multi-configurational expansion. The projector ensures orthogonality between orbital pairs which have Lagrange multipliers associated with them, by projecting out components of other orbitals every time the OrbitalHamiltonian action on orbital is computed.

source

Atoms.Projector — Type.

Projector(ϕs)

Represents the projector out of the subspace spanned by the radial orbitals ϕs

source

Atoms.projectout! — Function.

projectout!(y, projector)

Project out all components of y parallel to the radial orbitals projector.ϕs.

source

SCF.energy_matrix! — Method.

energy_matrix!(H, hamiltonian, ϕ)

Compute the contribution of hamiltonian to the Hamiltonian matrix H by repeatedly acting on the associated radial orbital ϕ with the different multi-configurational OrbitalHamiltonianTerms of hamiltonian.

source

Base.filter — Method.

filter(fun::Function, H::OrbitalHamiltonian)

Filter the OrbitalHamiltonianTerms of H according to the predicate fun.

source

Base.copyto! — Method.

copyto!(dest::AbstractMatix, hamiltonian::OrbitalHamiltonian)

Materialize the orbital hamiltonian into matrix form and store it in dest, using the current values of all other orbitals. This is only possible if the orbital hamiltonian does not contain any ExchangePotentials or SourceTerms, since the former is non-local (and thus not representable as a matrix) and the latter is not a linear operator (but an affine one).

Typical usage is to compute an easily factorizable matrix that can be used for preconditioning the solution of the full equation.

source

Base.:+ — Method.

h::OrbitalHamiltonian + λ::UniformScaling

Shift the OrbitalHamiltonian h by λ.

source

Base.:- — Method.

h::OrbitalHamiltonian - λ::UniformScaling

Shift the OrbitalHamiltonian h by -λ.

source

SCF.KrylovWrapper — Type.

SCF.KrylovWrapper(hamiltonian::OrbitalHamiltonian)

Construct a KrylovWrapper such that hamiltonian, that acts on function spaces, can be used in a Krylov solver, which works with linear algebra vector spaces.

source

LinearAlgebra.mul! — Method.

mul!(y, A::KrylovWrapper{T,<:OrbitalHamiltonian}, x)

Materialize the action of the OrbitalHamiltonian on the linear algebra vector x and store the result in y, by wrapping them both with the QuasiMatrix necessary to transform x and y to the function space of the Hamiltonian.

source

MatrixFactorizations.preconditioner — Function.

MatrixFactorizations.preconditioner(hamiltonian::OrbitalHamiltonian)

Return a factorization of the matrix corresponding to hamiltonian, where all terms arising from exchange and configuration interaction have been removes, since they cannot be represented by a matrix.

source

Orbital integrals and terms

Atoms.OrbitalIntegral — Type.

OrbitalIntegral{N}

Abstract type for integrals of rank N of orbitals, whose values need to be recomputed every time the orbitals are updated. Rank 0 corresponds to a scalar value, rank 1 to a diagonal matrix, etc.

source

Atoms.OrbitalOverlapIntegral — Type.

OrbitalOverlapIntegral(a, b, av, bv, value)

Represents the orbital overlap integral ⟨a|b⟩, for orbitals a and b, along with views of their radial orbitals av and bv and the current value of the integral.

source

SCF.update! — Method.

SCF.update!(oo::OrbitalOverlapIntegral)

Update the value of the integral oo.

source

Atoms.HFPotential — Type.

HFPotential(k, a, b, av, bv, V̂, poisson)

Represents the k:th multipole exansion of the Hartree–Fock potential formed by orbitals a and b (av and bv being views of their corresponding radial orbitals). V̂ is the resultant one-body potential formed, which can act on a third orbital and poisson computes the potential by solving Poisson's problem.

source

Atoms.DirectPotential — Type.

DirectPotential

Special case of HFPotential for the direct interaction, in which case the potential formed from two orbitals can be precomputed before acting on a third orbital.

source

SCF.update! — Method.

SCF.update!(p::DirectPotential)

Update the direct potential p by solving the Poisson problem with the current values of the orbitals forming the mutual density.

source

Base.Broadcast.materialize! — Method.

materialize!(::MulAdd{<:Any, <:Any, <:Any, T, <:AtomicOneBodyHamiltonian, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*H*x + β*y where H is a AtomicOneBodyHamiltonian and x and y are RadialOrbitals.

source

materialize!(ma::MulAdd{<:Any, <:Any, <:Any, T, <:DirectPotential, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*V̂*x + β*y where V̂ is a DirectPotential (with a precomputed direct potential computed via SCF.update!) and x and y are RadialOrbitals.

source

materialize!(ma::MulAdd{<:Any, <:Any, <:Any, T, <:ExchangePotential, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*V̂*x + β*y where V̂ is a ExchangePotential (by solving the Poisson problem with x as one of the constituent source orbitals in the mutual density) and x and y are RadialOrbitals.

source

Atoms.ExchangePotential — Type.

ExchangePotential

Special case of HFPotential for the exchange interaction, in which case the potential is formed from the orbital acted upon, along with another orbital, and then applied to a third orbital. Thus this potential cannot be precomputed, but must be recomputed every time the operator is applied. This makes this potential expensive to handle and the number of times it is applied should be minimized, if possible.

source

Base.Broadcast.materialize! — Method.

materialize!(::MulAdd{<:Any, <:Any, <:Any, T, <:AtomicOneBodyHamiltonian, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*H*x + β*y where H is a AtomicOneBodyHamiltonian and x and y are RadialOrbitals.

source

materialize!(ma::MulAdd{<:Any, <:Any, <:Any, T, <:DirectPotential, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*V̂*x + β*y where V̂ is a DirectPotential (with a precomputed direct potential computed via SCF.update!) and x and y are RadialOrbitals.

source

materialize!(ma::MulAdd{<:Any, <:Any, <:Any, T, <:ExchangePotential, Source, Dest})

Materialize the lazy multiplication–addition of the type y ← α*V̂*x + β*y where V̂ is a ExchangePotential (by solving the Poisson problem with x as one of the constituent source orbitals in the mutual density) and x and y are RadialOrbitals.

source

Atoms.SourceTerm — Type.

SourceTerm(operator, source_orbital, ov)

The point of SourceTerm is to implement inhomogeneous terms that contribute to the equation for an orbital, and whose input is some other source_orbital. This kind of term appears in multi-configurational problems.

source

Atoms.ShiftTerm — Type.

ShiftTerm(λ)

The point of ShiftTerm is to implement an overall energy shift of the Hamiltonian.

source

Orbital equations

AtomicOrbitalEquation
energy(hfeq::AtomicOrbitalEquation, term=:all)