Hydrogenic initialization

Hydrogenic initialization

Atoms.hydrogenic! — Function.

hydrogenic!(atom[; kwargs...])

Initialize the radial orbitals of atom to their unscreened hydrogenic values. This is done via simple diagonalization of the one-body Hamiltonian for each angular symmetry. The kwargs are passed on to diagonalize_one_body and can be used to influence how the diagonalization is performed.

source