Observables

Observables

Atoms.Observable — Type.

Observable

Represents a physical quantity that can be observed, which is calculated as the matrix element of an operator between two configurations. All physical observables are real.

source

Atoms.Observable — Method.

Observable(operator, atom, overlaps, integrals[; double_counted=false])

Construct an observable corresponding the operator acting on atom; if double_counted, only return those terms that would be double-counted, otherwise return the normal observable equations. overlaps is a list of non-orthogonal, integrals a list of common integrals, and integral_map is a mapping from symbolic integrals to OrbitalIntegrals.

source

Atoms.observe! — Function.

observe!(A::M, o::Observable)

Compute the observable o between all configurations and store the results as matrix elements of A.

source