Radial orbitals

Radial orbitals

Atoms.RadialOrbital — Type.

RadialOrbital

A radial orbital is represented using basis coupled with a vector of expansion coefficients with respect to that basis; the basis implemented as an AbstractQuasimatrix.

source

Atoms.RadialOrbitals — Type.

RadialOrbitals

A collection of radial orbitals is instead represented using a matrix of such expansion coefficients, where each matrix column corresponds to a single radial orbital.

source